Recreate the cached data sets for the ggRandomForests package

Recreate the cached data sets for the ggRandomForests package

cache_rfsrc_datasets(set = NA, save = TRUE, pth, ...)

Arguments

set

Defaults to all sets (NA), however for individual sets specify one or more of c("airq", "Boston", "iris", "mtcars", "pbc", "veteran")

save

Defaults to write files to the current data directory.

pth

the directory to store files.

...

extra arguments passed to randomForestSRC functions.

Details

Constructing random forests are computationally expensive, and the ggRandomForests operates directly on randomForestSRC objects. We cache computationally intensive randomForestSRC objects to improve the ggRandomForests examples, diagnostics and vignettes run times. The set of precompiled randomForestSRC objects are stored in the package data subfolder, however version changes in the dependent packages may break some functionality. This function was created to help the package developer deal with those changes. We make the function available to end users to create objects for further experimentation.

There are five cached data set types: '

rfsrc_data - rfsrc objects.

varsel_data - var.select minimal depth variable selection objects.

interaction_data - find.interaction minimal depth, pairwise variable interaction matrices.

partial_data - plot.variable objects (partial=TRUE) for partial variable dependence.

partial_coplot_data - plot.variable objects (partial=TRUE) for partial variable dependence.

For the following data sets: #'

_iris - The iris data set.

_airq - The airquality data set.

_mtcars - The mtcars data set.

_Boston - The Boston housing data set (MASS package).

_pbc - The pbc data set (randomForestSRC package).

_veteran - The veteran data set (randomForestSRC package).

See also