Predicted response data object

Extracts the predicted response values from the rfsrc object, and formats data for plotting the response using plot.gg_rfsrc.

Usage

# S3 method for class 'rfsrc'
gg_rfsrc(object, oob = TRUE, by, ...)

Arguments

object

rfsrc object

oob

boolean, should we return the oob prediction , or the full forest prediction.

by

stratifying variable in the training dataset, defaults to NULL

...

extra arguments

Value

gg_rfsrc object

Details

surv_type ("surv", "chf", "mortality", "hazard") for survival forests

oob boolean, should we return the oob prediction , or the full forest prediction.

See also