Find points evenly distributed along the vectors values.

This function finds point values from a vector argument to produce groups intervals. Setting groups=2 will return three values, the two end points, and one mid point (at the median value of the vector).

The output can be passed directly into the breaks argument of the cut function for creating groups for coplots.

quantile_pts(object, groups, intervals = FALSE)

Arguments

object

vector object of values.

groups

how many points do we want

intervals

should we return the raw points or intervals to be passed to the cut function

Value

vector of groups+1 cut point values.

See also

Examples

data(rfsrc_Boston)

# To create 6 intervals, we want 7 points.
# quantile_pts will find balanced intervals
rm_pts <- quantile_pts(rfsrc_Boston$xvar$rm, groups=6, intervals=TRUE)

# Use cut to create the intervals
rm_grp <- cut(rfsrc_Boston$xvar$rm, breaks=rm_pts)

summary(rm_grp)
#> (3.56,5.73] (5.73,5.98] (5.98,6.21] (6.21,6.47] (6.47,6.87] (6.87,8.78] 
#>          81          85          92          87          81          80